refactor(theme): split appearance rules resolve and tests modules
